@@ -1,10 +1,9 @@
-﻿namespace NChinese.Phonetic
+﻿namespace NChinese.Phonetic;
+
+// 中文字出現頻率
+public enum ChineseCharFrequence
 {
-    // 中文字出現頻率
-    public enum ChineseCharFrequence
-    {
-        Common = 0,             // 常用字
-        LessThanCommon = 1,     // 次常用字
-        Rare = 2                // 罕見字
-    }
+    Common = 0,             // 常用字
+    LessThanCommon = 1,     // 次常用字
+    Rare = 2                // 罕見字
 }
